keysmanager: clean up a little bit
this whole part of the code is absolutely chaotic, yet it's really
simple. I'm too lazy to re-write it right now. I started cleaning it up
a bit because tests didn't pass, and it was caused by a bug in the
underlying *salt* manager. I should have picked it up with git...