*: implement user password retrieval from cookie
We get the password from the user's cookie (it's encrypted using the
app's keys). Only from that can we decrypt his files. This is the safest
way I could come up with to store the password.

We don't want to store the user's password in our database (we store a
one way hash, but that's just for convenience. We can't get the original
password from that). If we encrypted the password using some key, then
if the attacker got access to the db, then he probably would also have
access to this key, and hence would get access to the user's password.
Instead, by storing the user's password in an encrypted cookie, the
attacker has to:

1. get the app's secret key (stored on the server)
2. get each individual user's cookie to get their password

The most efficient way for him to do get each individual cookie is to
inject his own code on the server, which would send every request to a
third party. That process takes time, and is hence easier to
detectable.